Add files via upload
Dataset has 2.2 Million rows

Concept:  Load a small batch of data

Concepts:  
Numerical features lend themselves to regression. 
Non-numerical features -- > classification

Concept:  Visualize the distribution of numerical data 

Distribution is somewhat informative when we load first 300 rows of data (which happens to be all Puerto Rico). 
When we visualize data from California, the graphs are not informative due to most of the data clustered in a small part of the range.  Graph axis is automatically according to the complete range.  That probably indicates there are some outliers, such as there is 1 point at 1 Million, where as rest of the data is in the range 1-10, so the automated visualization is not helpful.  

How to identify outliers?  Can we use clustering methods?  And then visualize the clusters on separate axes?